Add timespan filters to other aggregation, ensure it doesnt self-filter
